+++ src/pages/private/Library/Music/Albums/index.js
@@ -3,9 +3,9 @@ import { BackHandler } from 'react-native';
 import Program from '~/components/Program';
 import { PlayerContext } from '~/context/player.context';
 import api from '~/services/api';
-import { PlayList } from '../../PodCasts/Predications/styles';
+import { PlayList } from '../../Predication/Predications/styles';
 
-import { Container, Title } from './styles';
+import { Container } from './styles';
 
 export default function Albums() {
   const [newAlbums, setNewAlbums] = useState([]);
