+++ src/pages/private/Library/Predication/Predicator/index.js
