+++ src/services/config.dist.json
@@ -1,4 +1,4 @@
 {
   "urlOld": "http://192.168.1.12:3333",
-  "url": "https://test-api.koel.info-charisma.com/api"
+  "url": "https://api.ppd.editions-charisma.fr/api"
 }
